Sample improvements (#47)
* Use @AutoConfigureHttpGraphQlTester

* Polishing in controllers

There is no need to specify type name and field if those match the
source object type name and the method name respectively.

* Use ClassNameTypeResolver

We can use the built-in TypeResolver that matches class names to schema
type names. The resolver behavior can be customized further if needed.
See ClassNameTypeResolver javadoc for details.