Upgrade to Spring for GraphQL 1.3 and use built-in federation support (#80)
Spring for GraphQL 1.3 adds [built-in federation support](https://spring.io/blog/2024/02/21/spring-for-graphql-1-3-m1-released#apollo-federation) with annotated, controller methods. This PR contains the changes to switch to the built-in support. Also some minor refactoring.

I know 1.3 is not due until May 21, and Spring Boot 3.3 until May 23, but I wanted to have the changes ready and make sure all works as expected. For now this is based on release candidates.